Updates so that coded compiles with devkitArm r21.
This all compiles with r21. Sadly it doesn't all work.

git-svn-id: https://quirkysoft.googlecode.com/svn/trunk@455 89bda9af-7d2e-0410-a434-4dd01d8ab487

+++ libndspp/include/Background.h
@@ -195,7 +195,7 @@ namespace nds {
       } background_t;
 
       //! the background register to use
-      volatile unsigned int & m_DISPCNT;
+      volatile unsigned long int & m_DISPCNT;
       //! the current hardware state for this background
       background_t m_bg;
